+++ plom/baseMessenger.py
@@ -958,40 +958,74 @@ class BaseMessenger:
         finally:
             self.SRmutex.release()
 
-    def McreateRubric(self, new_rubric):
+    def McreateRubric(self, new_rubric: dict[str, Any]) -> dict[str, Any]:
         """Ask server to make a new rubric and get key back.
 
         Args:
-            new_rubric (dict): the new rubric info as dict.
+            new_rubric: the new rubric info as dict.  The server will
+                probably add other fields so you should not consider
+                this input data as canonical.  Instead, call back and
+                get the new rubric.
 
         Raises:
             PlomAuthenticationException: Authentication error.
             PlomSeriousException: Other error types, possible needs fix or debugging.
 
         Returns:
-            str: the key/id of the new rubric.
+            The dict key-value representation of the new rubric.  This is
+            generally not the same as the input data, for example, it has an
+            key/id.
         """
-        self.SRmutex.acquire()
+        with self.SRmutex:
+            try:
+                response = self.put(
+                    "/MK/rubric",
+                    json={
+                        "user": self.user,
+                        "token": self.token,
+                        "rubric": new_rubric,
+                    },
+                )
+                response.raise_for_status()
+                new_rubric = response.json()
+            except requests.HTTPError as e:
+                if response.status_code == 401:
+                    raise PlomAuthenticationException(response.reason) from None
+                if response.status_code == 406:
+                    raise PlomSeriousException(response.reason) from None
+                raise PlomSeriousException(
+                    f"Error when creating new rubric: {e}"
+                ) from None
+        if self.is_legacy_server():
+            # On legacy servers, `new_rubric` will actually just be the key
+            assert isinstance(new_rubric, str)
+            return self.get_one_rubric(new_rubric)
+        return new_rubric
+
+    def get_one_rubric(self, key: str) -> dict[str, Any]:
+        """Retrieve one rubric from its key.
+
+        I don't think we actually have an endpoint for this.  For now
+        we fake it by getting all rubrics and filtering.
+
+        Args:
+            The key/id of the rubric we want.
+
+        Raises:
+            PlomNoRubric: no such rubric.
+            PlomAuthenticationException: Authentication error.
+            PlomSeriousException: Other error types, possible needs fix or debugging.
+
+        Returns:
+            Dict representation of the rubric.
+        """
+        rubrics = self.MgetRubrics(None)
         try:
-            response = self.put(
-                "/MK/rubric",
-                json={
-                    "user": self.user,
-                    "token": self.token,
-                    "rubric": new_rubric,
-                },
-            )
-            response.raise_for_status()
-            new_key = response.json()
-            return new_key
-        except requests.HTTPError as e:
-            if response.status_code == 401:
-                raise PlomAuthenticationException(response.reason) from None
-            if response.status_code == 406:
-                raise PlomSeriousException(response.reason) from None
-            raise PlomSeriousException(f"Error when creating new rubric: {e}") from None
-        finally:
-            self.SRmutex.release()
+            # ensure there is exactly one matching rubric in each list and grab it
+            (r,) = [r for r in rubrics if r["id"] == key]
+        except ValueError:
+            raise PlomNoRubric(f"No rubric with key={key}") from None
+        return r
 
     def MgetRubrics(self, question: int | None = None) -> list[dict[str, Any]]:
         """Retrieve list of all rubrics from server for given question.
@@ -1004,17 +1038,17 @@ class BaseMessenger:
             PlomSeriousException: Other error types, possible needs fix or debugging.
 
         Returns:
-            list: list of dicts, possibly an empty list if server has no
-                rubrics for this question.
+            List of dicts, possibly an empty list if server has no
+            rubrics for this question.
         """
         if self.is_legacy_server():
             return self._legacy_getRubrics(question)
 
         with self.SRmutex:
             if question is None:
-                url = "/MK/rubric"
+                url = "/MK/rubrics"
             else:
-                url = f"/MK/rubric/{question}"
+                url = f"/MK/rubrics/{question}"
             try:
                 response = self.get_auth(url)
                 response.raise_for_status()
@@ -1051,17 +1085,23 @@ class BaseMessenger:
                     r["system_rubric"] = True
                 else:
                     r["system_rubric"] = False
+                # A special sentinel value for legacy server
+                # TODO: annoying b/c downstream needs to detect and not send to arrow
+                r.setdefault("last_modified", "unknown")
+                r.setdefault("modified_by_username", "")
             return rubrics
 
-    def MmodifyRubric(self, key, new_rubric):
+    def MmodifyRubric(self, key: str, new_rubric: dict[str, Any]) -> dict[str, Any]:
         """Ask server to modify a rubric and get key back.
 
         Args:
-            rubric (dict): the modified rubric info as dict.
+            key: the key/id of the rubric to change.
+            new_rubric: the changes we want to make as a key-value dict.
 
         Returns:
-            str: the key/id of the rubric.  Currently should be unchanged
-            from what you sent.  But this behaviour might change in the future.
+            The dict key-value representation of the new rubric.  This is
+            generally not the same as the input data, for example, it has an
+            key/id.  You should use this returned rubric and discard the input.
 
         Raises:
             PlomAuthenticationException: Authentication error.
@@ -1071,40 +1111,42 @@ class BaseMessenger:
             PlomConflict: two users try to modify the rubric.
             PlomSeriousException: Other error types, possible needs fix or debugging.
         """
-        self.SRmutex.acquire()
-        try:
-            response = self.patch(
-                f"/MK/rubric/{key}",
-                json={
-                    "user": self.user,
-                    "token": self.token,
-                    "rubric": new_rubric,
-                },
-            )
-            response.raise_for_status()
-            new_key = response.json()
-            return new_key
-        except requests.HTTPError as e:
-            if response.status_code == 401:
-                raise PlomAuthenticationException(response.reason) from None
-            elif response.status_code == 400:
-                raise PlomSeriousException(response.reason) from None
-            elif response.status_code == 403:
-                raise PlomNoPermission(response.reason) from None
-            elif response.status_code == 404:
-                raise PlomNoRubric(response.reason) from None
-            elif response.status_code == 406:
-                raise PlomInconsistentRubric(response.reason) from None
-            elif response.status_code == 409:
-                if self.is_legacy_server():
-                    # legacy sends 409 for not-found
+        with self.SRmutex:
+            try:
+                response = self.patch(
+                    f"/MK/rubric/{key}",
+                    json={
+                        "user": self.user,
+                        "token": self.token,
+                        "rubric": new_rubric,
+                    },
+                )
+                response.raise_for_status()
+                new_rubric = response.json()
+            except requests.HTTPError as e:
+                if response.status_code == 401:
+                    raise PlomAuthenticationException(response.reason) from None
+                elif response.status_code == 400:
+                    raise PlomSeriousException(response.reason) from None
+                elif response.status_code == 403:
+                    raise PlomNoPermission(response.reason) from None
+                elif response.status_code == 404:
                     raise PlomNoRubric(response.reason) from None
-                raise PlomConflict(response.reason) from None
-            raise PlomSeriousException(
-                f"Error of type {e} when creating new rubric"
-            ) from None
-        finally:
-            self.SRmutex.release()
+                elif response.status_code == 406:
+                    raise PlomInconsistentRubric(response.reason) from None
+                elif response.status_code == 409:
+                    if self.is_legacy_server():
+                        # legacy sends 409 for not-found
+                        raise PlomNoRubric(response.reason) from None
+                    raise PlomConflict(response.reason) from None
+                raise PlomSeriousException(
+                    f"Error of type {e} when creating new rubric"
+                ) from None
+        if self.is_legacy_server():
+            # On legacy servers, `new_rubric` will actually just be the key
+            assert isinstance(new_rubric, str)
+            return self.get_one_rubric(new_rubric)
+        return new_rubric
 
     def get_pagedata(self, code):
         """Get metadata about the images in this paper."""