+++ plom/client/rubric_add_dialog.py
@@ -12,6 +12,8 @@ import re
 import sys
 from typing import Any
 
+import arrow
+
 if sys.version_info >= (3, 9):
     from importlib import resources
 else:
@@ -161,7 +163,8 @@ class AddRubricBox(QDialog):
 
         Args:
             parent (QWidget): the parent window.
-            username (str)
+            username (str): who is creating this rubric or who is
+                modifying this rubric.
             maxMark (int)
             question_number (int)
             question_label (str)
@@ -171,7 +174,6 @@ class AddRubricBox(QDialog):
                 Otherwise, this has the current comment data.
 
         Keyword Args:
-            annotator_size (QSize/None): size of the parent annotator
             groups (list): optional list of existing/recommended group
                 names that the rubric could be added to.
             add_to_group (str/None): preselect this group in the scope
@@ -191,8 +193,13 @@ class AddRubricBox(QDialog):
         self.question_number = question_number
         self.version = version
         self.maxver = maxver
+        self._username = username
 
+        self._is_edit = False
         if com:
+            self._is_edit = True
+
+        if self.is_edit():
             self.setWindowTitle("Modify rubric")
         else:
             self.setWindowTitle("Add new rubric")
@@ -205,7 +212,7 @@ class AddRubricBox(QDialog):
         self.TEmeta = WideTextEdit()
         # cannot edit these
         self.label_rubric_id = QLabel("Will be auto-assigned")
-        self.Luser = QLabel()
+        self.last_modified_label = QLabel()
 
         sizePolicy = QSizePolicy(
             QSizePolicy.Policy.Expanding, QSizePolicy.Policy.MinimumExpanding
@@ -427,7 +434,7 @@ class AddRubricBox(QDialog):
         flay.addRow("Meta", self.TEmeta)
 
         flay.addRow("Rubric ID", self.label_rubric_id)
-        flay.addRow("Created by", self.Luser)
+        flay.addRow("", self.last_modified_label)
 
         buttons = QDialogButtonBox(
             QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel
@@ -473,9 +480,20 @@ class AddRubricBox(QDialog):
                     self.typeRB_absolute.setChecked(True)
                 else:
                     raise RuntimeError(f"unexpected kind in {com}")
-            if com["id"]:
+            if com.get("id"):
                 self.label_rubric_id.setText(str(com["id"]))
-            self.Luser.setText(com.get("username", ""))
+            s = f'Created by {com.get("username", "unknown")}'
+            lastmod = com.get("last_modified", "unknown")
+            # Note sure if would be None but seems harmless (or no more harmful
+            # than "unknown" sentintel from legacy anyway)
+            if lastmod is not None and lastmod != "unknown":
+                rev = com.get("revision", 0)
+                s += f", revision {rev}"
+                if rev > 0:
+                    s += f", last modified {arrow.get(lastmod).humanize()}"
+                    s += f' by {com["modified_by_username"]}'
+            self.last_modified_label.setText(s)
+            self.last_modified_label.setWordWrap(True)
             if com.get("versions"):
                 self.version_specific_cb.setChecked(True)
                 self.version_specific_le.setText(
@@ -541,7 +559,9 @@ class AddRubricBox(QDialog):
                 "Notes about this rubric such as hints on when to use it.\n\n"
                 "Not shown to student!"
             )
-            self.Luser.setText(username)
+            self.last_modified_label.setText(
+                f"You ({self._username}) are creating a new rubric"
+            )
             if add_to_group:
                 assert add_to_group in groups, f"{add_to_group} not in groups={groups}"
                 self.group_checkbox.setChecked(True)
@@ -550,6 +570,10 @@ class AddRubricBox(QDialog):
                 self.scopeButton.animateClick()
         self.subsRemakeGridUI(params)
         self.hiliter.setSubs([x for x, _ in params])
+
+    def is_edit(self):
+        """Answer true if we are editing a rubric (rather than making a new one)."""
+        return self._is_edit
 
     def subsMakeGridUI(self, params):
         maxver = self.maxver
@@ -851,9 +875,6 @@ class AddRubricBox(QDialog):
             display_delta = f"{value} of {out_of}"
         else:
             raise RuntimeError("no radio was checked")
-        username = self.Luser.text().strip()
-        # only meaningful if we're modifying
-        rubricID = self.label_rubric_id.text().strip()
 
         vers = self.get_versions_list()
 
@@ -862,7 +883,6 @@ class AddRubricBox(QDialog):
         rubric = self._old_rubric
         rubric.update(
             {
-                "id": rubricID,
                 "kind": kind,
                 "display_delta": display_delta,
                 "value": value,
@@ -870,10 +890,16 @@ class AddRubricBox(QDialog):
                 "text": txt,
                 "tags": tags,
                 "meta": meta,
-                "username": username,
                 "question": self.question_number,
                 "versions": vers,
                 "parameters": params,
             }
         )
+        if not self.is_edit():
+            rubric.update(
+                {
+                    "username": self._username,
+                }
+            )
+
         return rubric