+++ plom_server/Preparation/views/pqv_mapping.py
@@ -83,6 +83,7 @@ class PQVMappingDownloadView(ManagerRequiredView):
 class PQVMappingDeleteView(ManagerRequiredView):
     def delete(self, request: HttpRequest) -> HttpResponse:
         pqvs = PQVMappingService()
+        # fails with ValueError if state disallows delete
         pqvs.remove_pqv_map()
         return HttpResponseClientRedirect(".")
 
@@ -101,6 +102,7 @@ class PQVMappingView(ManagerRequiredView):
             "pqv_mapping_present": pqvs.is_there_a_pqv_map(),
             "number_of_students": sss.how_many_students(),
             "student_list_present": sss.are_there_students(),
+            "have_papers_been_printed": PapersPrinted.have_papers_been_printed(),
         }
 
         prenamed_papers_list = list(
@@ -132,8 +134,6 @@ class PQVMappingView(ManagerRequiredView):
         return context
 
     def get(self, request: HttpRequest) -> HttpResponse:
-        if PapersPrinted.have_papers_been_printed():
-            return redirect("prep_qvmapping_view")
         context = self.build_context()
         return render(request, "Preparation/pqv_mapping_manage.html", context)
 
@@ -160,23 +160,3 @@ class PQVMappingView(ManagerRequiredView):
         pqvs = PQVMappingService()
         pqvs.generate_and_set_pqvmap(number_to_produce, first=first)
         return HttpResponseRedirect(".")
-
-
-class PQVMappingReadOnlyView(ManagerRequiredView):
-    def build_context(self):
-        context = super().build_context()
-        pqvs = PQVMappingService()
-        pss = PrenameSettingService()
-
-        context.update(
-            {
-                "prenaming": pss.get_prenaming_setting(),
-                "question_labels_html": SpecificationService.get_question_html_label_triples(),
-                "pqv_table": pqvs.get_pqv_map_as_table(pss.get_prenaming_setting()),
-            }
-        )
-        return context
-
-    def get(self, request: HttpRequest) -> HttpResponse:
-        context = self.build_context()
-        return render(request, "Preparation/pqv_mapping_view.html", context)