+++ plom_server/Profile/edit_profile_form.py
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 # Copyright (C) 2022 Brennen Chiu
-# Copyright (C) 2023 Colin B. Macdonald
+# Copyright (C) 2023-2024 Colin B. Macdonald
 
 from django.contrib.auth.forms import UserChangeForm
 from django.contrib.auth.models import User
@@ -9,4 +9,5 @@ from django.contrib.auth.models import User
 class EditProfileForm(UserChangeForm):
     class Meta:
         model = User
+        # Note we only use first_name; last_name field from User unused
         fields = ["first_name", "email"]