+++ CHANGELOG.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+* Rubrics now keep more careful records about who modified them and when.
 
 ### Fixed
 * Legacy workflow: `plom-scan upload ./foo.pdf` will now explicitly fail informing you do keep path info out of bundle names.
 * Processing PDF scans: don't complain about rounding errors when instead some metadata mixup about width/height has occurred.
+* Client: clicking "sync" to synchronous rubrics from sever no longer resets selection.
 
 
 ## [0.15.2] - 2024-03-11